requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
streamlit>=1.28.0
tavily-python>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
arxiv>=2.1.0
typing-extensions>=4.8.0

